ENH: refactored factory data.  Cleaned transforms a bit
