+++ data/__init__.py
@@ -1,4 +1,2 @@
-from .factory_ref.reference_tables import (
-    countries as factory_countries, 
-    milestones as factory_milestones
-    )
+from .factory_ref.reference_countries import countries as factory_countries
+from .factory_ref.reference_milestones import milestones as factory_milestones
