+++ data/factory_ref/reference_countries.py
@@ -1,73 +1,6 @@
 """
 Factory tables for loading database with ground truth or set categorical data.
 """
-
-# Milestone Information
-#   milestone_id => category_num+milestone_in_category_num
-milestones = (
-    {
-        'milestone_id': 11,
-        'name': 'discovery_started',
-        'category': 'pre-clinical',
-    },
-    {
-        'milestone_id': 12,
-        'name': 'pre_clinical_studies_started',
-        'category': 'pre-clinical',
-    },
-    {
-        'milestone_id': 13,
-        'name': 'lead_selection_finalized',
-        'category': 'pre-clinical'
-    },
-    {
-        'milestone_id': 21,
-        'name': 'clinical_batch_finalized',
-        'category': 'manufacturing',
-    },
-    {
-        'milestone_id': 31,
-        'name': 'ind_finalized',
-        'category': 'regulatory',
-    },
-    {
-        'milestone_id': 41,
-        'name': 'phase_1_started',
-        'category': 'clinical_development',
-    },
-    {
-        'milestone_id': 42,
-        'name': 'phase_2_started',
-        'category': 'clinical_development',
-    },
-    {
-        'milestone_id': 43,
-        'name': 'phase_3_started,
-        'category': 'clinical_development',
-    },
-    {
-        'milestone_id': 44,
-        'name': 'phase_4_started',
-        'category': 'clinical_development',
-    },
-    {
-        'milestone_id': 51,
-        'name': 'nda_finalized',
-        'category': 'regulatory',
-    },
-    {
-        'milestone_id': 61,
-        'name': 'completed',
-        'category': 'completion',
-    },
-    {
-        'milestone_id': 62,
-        'name': 'approved',
-        'category': 'completion',
-    },
-)
-
-
 ## Adapted from https://gist.github.com/Keeguon/2310008
 ## Special thanks to: Félix Bellanger keeguon
